SP-359 refactor: request / response 디렉터리 구조 변경
