+++ src/main/java/com/ludo/study/studymatchingplatform/study/service/dto/response/study/participant/ParticipantResponse.java
@@ -1,6 +1,6 @@
 package com.ludo.study.studymatchingplatform.study.service.dto.response.study.participant;
 
-import com.ludo.study.studymatchingplatform.study.domain.study.Participant;
+import com.ludo.study.studymatchingplatform.study.domain.study.participant.Participant;
 import com.ludo.study.studymatchingplatform.user.service.dto.response.UserResponse;
 
 import lombok.Builder;
